feat: Clean up
