+++ Cargo.toml
@@ -23,7 +23,8 @@ path = "src/bin/main.rs"
 required-features = ["cli"]
 
 [features]
-default = []
+# default = []
+default = ["permissive"]
 # default = ["permissive", "json_schema"]
 cli = ["clap"]
 json_schema = ["schemars", "serde_json"]
@@ -39,7 +40,7 @@ schemars = { version = "0.8.21", optional = true }
 regex = "1.10.5"
 url = { version = "2.5.2", features = ["serde"] }
 concat-idents = "1.1.5"
-struct-patch = { version = "0.6.0", path = "../../external/struct-patch/struct-patch", default-features = false }
+struct-patch = { version = "0.6.0", path = "../../external/struct-patch/struct-patch", default-features = false, features = ["option"] }
 
 [dev-dependencies]
 pretty_assertions_sorted = "1.2.3"
